+++ src/pretix/base/services/quotas.py
@@ -30,7 +30,7 @@ def refresh_quota_caches():
     ).filter(
         Q(cached_availability_time__isnull=True) |
         Q(cached_availability_time__lt=F('last_activity')) |
-        Q(cached_availability_time__lt=now() - timedelta(days=1), last_activity__gt=now() - timedelta(days=30))
+        Q(cached_availability_time__lt=now() - timedelta(hours=2), last_activity__gt=now() - timedelta(days=7))
     )
     for q in quotas:
         q.availability()
